Rename `rustc_interface/src/errors.rs` into `rustc_interface/src/diagnostics.rs`
