+++ compiler/rustc_interface/src/diagnostics.rs
