+++ compiler/rustc_interface/src/util.rs
@@ -30,7 +30,7 @@ use rustc_span::{SessionGlobals, Symbol, sym};
 use rustc_target::spec::Target;
 use tracing::info;
 
-use crate::errors;
+use crate::diagnostics;
 use crate::passes::parse_crate_name;
 
 /// Function pointer type that constructs a new CodegenBackend.
@@ -90,12 +90,14 @@ pub(crate) fn check_abi_required_features(sess: &Session) {
 
     for feature in abi_feature_constraints.required {
         if !sess.unstable_target_features.contains(&Symbol::intern(feature)) {
-            sess.dcx().emit_warn(errors::AbiRequiredTargetFeature { feature, enabled: "enabled" });
+            sess.dcx()
+                .emit_warn(diagnostics::AbiRequiredTargetFeature { feature, enabled: "enabled" });
         }
     }
     for feature in abi_feature_constraints.incompatible {
         if sess.unstable_target_features.contains(&Symbol::intern(feature)) {
-            sess.dcx().emit_warn(errors::AbiRequiredTargetFeature { feature, enabled: "disabled" });
+            sess.dcx()
+                .emit_warn(diagnostics::AbiRequiredTargetFeature { feature, enabled: "disabled" });
         }
     }
 }
@@ -609,7 +611,7 @@ pub fn build_output_filenames(attrs: &[ast::Attribute], sess: &Session) -> Outpu
         &sess.opts.output_types,
         sess.io.output_file == Some(OutFileName::Stdout),
     ) {
-        sess.dcx().emit_fatal(errors::MultipleOutputTypesToStdout);
+        sess.dcx().emit_fatal(diagnostics::MultipleOutputTypesToStdout);
     }
 
     let crate_name =
@@ -650,16 +652,16 @@ pub fn build_output_filenames(attrs: &[ast::Attribute], sess: &Session) -> Outpu
             let unnamed_output_types =
                 sess.opts.output_types.values().filter(|a| a.is_none()).count();
             let ofile = if unnamed_output_types > 1 {
-                sess.dcx().emit_warn(errors::MultipleOutputTypesAdaption);
+                sess.dcx().emit_warn(diagnostics::MultipleOutputTypesAdaption);
                 None
             } else {
                 if !sess.opts.cg.extra_filename.is_empty() {
-                    sess.dcx().emit_warn(errors::IgnoringExtraFilename);
+                    sess.dcx().emit_warn(diagnostics::IgnoringExtraFilename);
                 }
                 Some(out_file.clone())
             };
             if sess.io.output_dir.is_some() {
-                sess.dcx().emit_warn(errors::IgnoringOutDir);
+                sess.dcx().emit_warn(diagnostics::IgnoringOutDir);
             }
 
             let out_filestem =
